+++ jfxui/src/uiTest/java/org/itsallcode/whiterabbit/jfxui/testutil/model/MonthlyProjectReportWindow.java
@@ -5,18 +5,22 @@ import static org.assertj.core.api.Assertions.assertThat;
 import java.time.Duration;
 
 import org.itsallcode.whiterabbit.api.model.Project;
+import org.itsallcode.whiterabbit.jfxui.testutil.TestUtil;
+import org.itsallcode.whiterabbit.jfxui.ui.MonthlyProjectReportViewer;
 import org.itsallcode.whiterabbit.jfxui.ui.MonthlyProjectReportViewer.ReportRow;
 import org.testfx.api.FxRobot;
 import org.testfx.assertions.api.Assertions;
 
 import javafx.scene.input.KeyCode;
+import javafx.stage.Stage;
 import javafx.stage.Window;
 
 public class MonthlyProjectReportWindow
 {
     private final FxRobot robot;
     private final Window window;
-    private final JavaFxTable<ReportRow> table;
+
+    private final JavaFxTable<MonthlyProjectReportViewer.ReportRow> table;
 
     MonthlyProjectReportWindow(final FxRobot robot, final Window window)
     {
@@ -36,6 +40,23 @@ public class MonthlyProjectReportWindow
     {
         robot.type(KeyCode.ESCAPE);
         Assertions.assertThat(window).isNotShowing();
+    }
+
+    public void gotoPreviousMonth()
+    {
+        robot.clickOn("#prev-month-button");
+    }
+
+    public void gotoNextMonth()
+    {
+        robot.clickOn("#next-month-button");
+    }
+
+    public MonthlyProjectReportWindow assertWindowTitle(final String expectedTitle)
+    {
+        final String title = ((Stage) window).getTitle();
+        TestUtil.retryAssertion(Duration.ofSeconds(5), () -> assertThat(title).isEqualTo(expectedTitle));
+        return this;
     }
 
     public MonthlyProjectReportWindow assertRowCount(final int expectedRowCount)